requirements.txt
async-timeout==4.0.3
certifi==2023.7.22
charset-normalizer==3.2.0
filelock==3.12.3
grpcio==1.57.0
grpcio-tools==1.57.0
idna==3.4
Jinja2==3.1.2
joblib==1.4.2
MarkupSafe==2.1.3
mpmath==1.3.0
msgpack==1.0.7
networkx==3.1
numpy==1.25.2
opencv-contrib-python-headless==4.5.4.60
opencv-python-headless==4.5.4.60
pandas==2.0.3
Pillow==10.0.0
protobuf==4.24.1
psutil==7.0.0
psycopg2-binary==2.9.7
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2023.3
redis==5.0.0
requests==2.31.0
scikit-learn==1.6.1
scipy==1.15.2
six==1.16.0
sympy==1.12
threadpoolctl==3.5.0
torch==1.11.0
torch-vision==0.1.6.dev0
torchvision==0.12.0
tqdm==4.66.1
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
